feat(evaluate-plugin): add /evaluate:context-engineering audit skill
Moves the Channel M scanner into the skill that owns it and adds the
skill itself. scripts/check-context-engineering.py becomes a shim so
`just lint-context-engineering`, pre-commit, and CI keep working against
one implementation rather than two — the C4 single-source-of-truth
property the scanner measures.

The move surfaced a real bug the relocation caused: the default
--project-dir was derived from __file__, so once the scanner lived
inside a skill it scanned the skill's own directory and reported
SKILL_COUNT=0 — a silent "nothing to fix" indistinguishable from a clean
repo. The default now walks up from the CWD to the repo root, and the
self-test asserts a bare invocation from a fixture root finds it.

The skill keeps Channel M and Channel J separate rather than averaging
them, ranks findings by always-loaded cost rather than raw size, and
requires at least one counter-finding — a unit that earns its tokens —
so a cut-everything report cannot pass as a clean result.

Also adds the benchmark's reproduction tooling (quote_check.py,
synthesize.py), which the prior benchmark ran from a scratchpad and
never committed.
