+++ docs/benchmarks/2026-07-context-engineering/judgments/claude-md_j1.json
@@ -1,0 +1,35 @@
+{
+  "unit": "CLAUDE.md",
+  "unit_kind": "always_loaded",
+  "judge_id": "j1",
+  "dimensions": {
+    "C1": {"score": 3, "rationale": "Constraint density is low and the few hard directives concern real contracts or irreversible steps and state their reason; residual style prescriptions in the Conventions section ('use tables', 'include both short and long flag forms') are phrased as prohibitions rather than delegated judgment, which keeps it from 4.", "evidence": []},
+    "C2": {"score": "n/a", "rationale": "A memory file with no invocable interface: no arguments, flags, or parameter surface to make self-describing.", "evidence": []},
+    "C3": {"score": 4, "rationale": "Detail is deferred aggressively and by name: a 38-row index of separately-loadable rule files plus explicit 'see X' pointers to skill-development.md, agentic-optimization.md, docs/pi-export.md, adapters/README.md and docs/blueprint/README.md; short of 5 because rare reference detail (marketplace entry JSON, release-please-config block, manifest snippet) is still inlined in the entry point rather than split out.", "evidence": [{"quote": "See `.claude/rules/skill-development.md` for detailed patterns."}, {"quote": "| `.claude/rules/hooks-reference.md` | Complete hook event reference (2.1.50+): all events, schemas, timeouts, PermissionRequest |"}]},
+    "C4": {"score": 2, "rationale": "Several sections restate guidance that already lives in an auto-injected rule file and then point at that same file: the Agentic Optimization bullets duplicate agentic-optimization.md's compact-output/fail-fast/CI-mode principles, and the Git Workflow section reproduces the format, type list and worked example from conventional-commits.md before deferring to it; the naming bullets overlap skill-naming.md. The excellent rules index shows the file knows the pointer-only pattern, which is why the copies stand out.", "evidence": [{"quote": "- **Compact output**: `--dots`, `--reporter=github`, `-c`"}, {"quote": "See `.claude/rules/agentic-optimization.md` for detailed patterns."}]},
+    "C5": {"score": 2, "rationale": "Most of the file fails the every-turn test: an ASCII tree of a layout visible from the repo, a Plugin Organization table, and four multi-step procedures (skill creation, user-invocable skill creation, plugin lifecycle create/delete with inline JSON, development workflow) that only matter once a user asks for that specific task and are exactly the material the rubric says to promote to a skill and reference by name. Real gotchas are present (blueprint disabled-task reasons, autonomy-level-3 consumer-only, the pi tier-vs-adapter successor state, the deliberately non-matching blueprint commit scope) but occupy a minority of the tokens; two orphaned rule rows appended after the Conventions bullet list show the surface is large enough to drift.", "evidence": [{"quote": "1. Create skill directory: `mkdir -p <plugin>/skills/<skill-name>`"}, {"quote": "When creating, modifying, or deleting a plugin, update these files:"}, {"quote": "1. **Research documentation** - Use context7, web search"}, {"quote": "| `.claude/rules/drift-detection-triggering.md` | drift detection triggering |"}]},
+    "C6": {"score": 4, "rationale": "Where work is mechanical the file names the artifact that owns it rather than describing the check in prose: a YAML classification file as source of truth with its enforcing script, CI drift checks by script path, and just recipes as the runnable entry points; short of 5 because it rarely says whether to run or read the artifact, and the plugin-creation config is spelled out as literal snippets where a scaffolding script or schema could state it.", "evidence": [{"quote": "`pi/tiers.yaml` (tier classification, enforced by `scripts/check-pi-tiers.sh`)"}, {"quote": "The `Plugin: Enablement drift` check (`scripts/check-enabled-plugins-drift.sh`) fails CI if a marketplace plugin is left disabled."}]}
+  },
+  "top_finding": "Promote the four always-loaded procedures (Creating New Skills, Creating User-Invocable Skills, Plugin Lifecycle, Development Workflow) into an authoring skill and leave one named reference line each, so the always-loaded surface keeps only the gotchas.",
+  "counter_finding": "The Blueprint constrained-dogfooding section earns its tokens outright: the disabled-task reasons, the consumer-only autonomy-level-3 caveat and the deliberately non-matching `blueprint` commit scope are non-discoverable traps that would cause real damage if a model re-derived them from the repo.",
+  "section_verdicts": [
+    {"section": "# claude-plugins", "verdict": "keep", "why": "one-line statement of what the repo is, exactly the lightweight framing the anchor asks for"},
+    {"section": "## Project Structure", "verdict": "drop", "why": "an ASCII tree of a layout any reader sees from the filesystem"},
+    {"section": "## Rules", "verdict": "lean", "why": "a valuable named index, but several row descriptions have grown into paragraph summaries and two rows have drifted outside the table"},
+    {"section": "## Creating New Skills", "verdict": "promote-to-skill", "why": "a scaffolding procedure that only matters once a skill is actually being authored"},
+    {"section": "## Creating User-Invocable Skills", "verdict": "promote-to-skill", "why": "same task-triggered frontmatter procedure, split across two sections"},
+    {"section": "## Agentic Optimization", "verdict": "drop", "why": "restates agentic-optimization.md and then points at it"},
+    {"section": "## Plugin Organization", "verdict": "drop", "why": "taxonomy inferable from the plugin directory names"},
+    {"section": "## Plugin Lifecycle", "verdict": "promote-to-skill", "why": "long create/delete checklist with inline JSON, needed only when a plugin is added or removed"},
+    {"section": "## Git Workflow", "verdict": "lean", "why": "the automation contract is worth one line; the format spec, type list and worked example belong solely to conventional-commits.md"},
+    {"section": "## Development Workflow", "verdict": "drop", "why": "a generic seven-step process the model would follow anyway"},
+    {"section": "## Local-model export", "verdict": "lean", "why": "the tier-vs-adapter successor state is a genuine gotcha; the recipe inventory belongs in the linked docs"},
+    {"section": "## Blueprint (constrained dogfooding)", "verdict": "keep", "why": "non-discoverable constraints whose violation causes real damage"},
+    {"section": "## Conventions", "verdict": "lean", "why": "naming rules overlap skill-naming.md and the table/flag bullets are style prescriptions"}
+  ],
+  "leakage": {
+    "house_standards_seen": ["CLAUDE.md itself is auto-injected into every agent context in this repo, so the unit under judgment was already present before I read it", "several .claude/rules/*.md are auto-injected too (pr-branch-sync, loop-integrity, bash-tool-replacements, gh-json-fields, parallel-safe-queries, terminology, conventional-commits, task-id-stability, offload-to-deterministic-substrate, docs-currency, agent-coworker-detection, handling-blocked-hooks)", "the excluded house conventions (conventional-commit scopes, release-please dogfooding, plugin lifecycle bookkeeping) are visible in the injected copy"],
+    "influenced_scoring": false,
+    "note": "I re-read CLAUDE.md from disk and scored that file; all quotes were taken from the on-disk read. The injected rule files were used only as C4 evidence that specific guidance already has another owner (agentic-optimization.md, conventional-commits.md, skill-naming.md), which the rubric explicitly permits judges to discover. I did not score the desirability of any excluded house convention: the Git Workflow and Plugin Lifecycle sections were assessed only for single-sourcing and always-loaded cost, never for whether the conventions themselves are correct. I did not open any of the prohibited files."
+  }
+}
