+++ docs/benchmarks/2026-07-context-engineering/judgments/comfy-node_j1.json
@@ -1,0 +1,16 @@
+{
+  "unit": "comfyui-plugin/skills/comfy-node/SKILL.md",
+  "unit_kind": "skill",
+  "judge_id": "j1",
+  "dimensions": {
+    "C1": {"score": 5, "rationale": "Nearly every hard constraint is load-bearing (irreversible shared-infra apply, a red pack reaching a remote, a hook workaround that costs a rename/default-branch recovery) and states its own rationale, so the model can generalise beyond the enumerated case; naming, variant choice, and topics are delegated to judgment via a derive-and-confirm table rather than prohibited.", "evidence": [{"quote": "(merging the gitops PR) is intentionally human — it feeds the apply pipeline on"}, {"quote": "GitHub App *cannot* create repos on user accounts — that is exactly why the"}]},
+    "C2": {"score": 4, "rationale": "The scaffold parameter surface is declared as a table whose alternatives are enumerated at the point of declaration with selection criteria, and an invalid combination is named inline ('modal variants only; omit for gesture'); it falls short of 5 because the skill's own invocation surface is undeclared free text and ~20 command blocks are one fully-instantiated worked example with no stated placeholder convention.", "evidence": [{"quote": "| `--variant` | `gesture` for canvas interactions (resize/move/region); `frontend` for a per-widget modal; `backend` only if it reads disk / serves data. | `gesture` |"}]},
+    "C3": {"score": 3, "rationale": "The heaviest material is genuinely deferred and named — the generator runs as scaffold.py without being read into context, the screenshot pipeline and files-only path are handed to sibling skills by name, and the apply pipeline to gitops/CLAUDE.md — but the skill owns no supporting file of its own, so its rarer paths (failure modes, the import-block-removal follow-up, notes) all sit inline in one flat body; competent single-level disclosure, not multi-file.", "evidence": []},
+    "C4": {"score": 2, "rationale": "The gitops-adoption procedure here is duplicated almost wholesale by the sibling skill foundryvtt-plugin/skills/foundryvtt-module/SKILL.md (same phase structure, same repositories.tf/main.tf import-block edits, same branch-protection hook note, same failure-mode rows) with no reference in either direction, and the two copies have already drifted; the unit also restates its own Phase 2 finishing-pass paragraph as a Notes bullet rather than a pointer. Credit for the two clean ownership boundaries it does draw (comfyui-node-scaffold owns file generation, gitops/CLAUDE.md owns the apply chain) keeps it above 1.", "evidence": [{"quote": "- The scaffold now emits the registry finishing-pass pieces (icon/banner SVGs +"}]},
+    "C5": {"score": "n/a", "rationale": "This is a SKILL.md body, which costs tokens only once the skill fires; the always-loaded budget dimension applies to CLAUDE.md and .claude/rules/* per the rubric's scope caveat.", "evidence": []},
+    "C6": {"score": 4, "rationale": "Mechanical work is consistently handed to artifacts rather than described: an executable generator invoked by exact command line, exact HCL blocks to insert instead of prose about what to add, deterministic preflight one-liners emitting free/EXISTS, and a runnable green gate before anything reaches a remote. Short of 5 because the Phase 5 wiring verification is asserted in prose about expected output rather than an exit-code gate, and the orchestration ships no self-verifying script or fixture of its own.", "evidence": [{"quote": "to = github_repository.this[\"comfyui-touch-resize\"]"}, {"quote": "`just check` must pass before anything is pushed. If it fails, fix locally and"}]}
+  },
+  "top_finding": "Designate one owner for the shared gitops repo-adoption procedure (Phases 3-5, the branch-protection note, the failure-mode rows) and have this skill reference it rather than carry its own copy, since the twin in foundryvtt-plugin/skills/foundryvtt-module/SKILL.md has already drifted from it.",
+  "counter_finding": "The Phase 3 explanation of why main is seeded directly, plus the branch-protection hook note, earn their tokens outright: both pre-empt a non-obvious and expensive rename/default-branch/base-branch recovery that a model reasoning from ordinary feature-branch habits would walk straight into.",
+  "leakage": {"house_standards_seen": ["CLAUDE.md (project instructions, incl. Plugin Lifecycle and Conventions)", ".claude/rules/skill-quality.md (auto-injected despite being on the exclusion list)", ".claude/rules/skill-development.md (auto-injected, on exclusion list)", ".claude/rules/skill-evaluation.md (auto-injected, on exclusion list)", ".claude/rules/skill-consolidation.md", ".claude/rules/skill-execution-structure.md", ".claude/rules/skill-argument-handling.md", ".claude/rules/skill-naming.md", ".claude/rules/skill-fork-context.md", ".claude/rules/agentic-optimization.md", ".claude/rules/agentic-permissions.md", ".claude/rules/auto-mode.md", ".claude/rules/regression-testing.md", ".claude/rules/conventional-commits.md", ".claude/rules/offload-to-deterministic-substrate.md", ".claude/rules/version-pinning.md", ".claude/rules/sandbox-guidance.md", ".claude/rules/loop-integrity.md", ".claude/rules/pr-branch-sync.md", ".claude/rules/agent-coworker-detection.md", ".claude/rules/bash-tool-replacements.md", ".claude/rules/gh-json-fields.md", ".claude/rules/parallel-safe-queries.md", ".claude/rules/task-id-stability.md", ".claude/rules/terminology.md", ".claude/rules/docs-currency.md", ".claude/rules/handling-blocked-hooks.md"], "influenced_scoring": false, "note": "Several files on the hard-constraint exclusion list (skill-quality.md, skill-development.md, skill-evaluation.md) were auto-injected into context by the repo before I could avoid them; I did not open them deliberately and did not score against any of their criteria. Specifically: I did not credit or penalise the '## When to Use This Skill' table, the absent 'Agentic Optimizations' table, the 13,053-char body against any char ceiling, the absence of a file literally named REFERENCE.md, description length, the created/modified/reviewed dates, or the model: opus setting. C3 was scored on disclosure structure (what is deferred and how it is named), not on size. skill-consolidation.md's cross-plugin dedupe guidance was visible, but the C4 finding rests on the rubric's own 'two sibling skills' defect definition and on drift I verified directly by reading the foundryvtt twin."}
+}
