+++ docs/benchmarks/2026-07-context-engineering/judgments/git-commit_j1.json
@@ -1,0 +1,16 @@
+{
+  "unit": "git-plugin/skills/git-commit/SKILL.md",
+  "unit_kind": "skill",
+  "judge_id": "j1",
+  "dimensions": {
+    "C1": {"score": 4, "rationale": "The hard constraints are load-bearing and state why the violation is a real bug (heredoc escaping corrupting git log and needing an --amend; the pr-sync verdict gate before adding commits to a merged branch), and softer guidance is framed as preference rather than prohibition ('Explicit staging (preferred)'). It falls short of 5 only because the trailing Best Practices list still spends tokens on style defaults a capable model already applies (imperative mood, 72-char subject, 'Always reference issues').", "evidence": [{"quote": "Inside a quoted heredoc (`<<'EOF'`), backticks, `$`, and `\\` are already literal — never backslash-escape them."}, {"quote": "**Explicit staging** (preferred):"}]},
+    "C2": {"score": 4, "rationale": "The skill's operational surface is defined by enumeration rather than by accumulated examples: commit types are a closed table with a stated effect column (version bump), and issue-reference keywords are enumerated with their exact semantics at the point of declaration, so the model can pick without reading usage samples. Not 5 because there is no declared parameter surface at all (no args/argument-hint; user-invocable: false), and invocation is still taught by a body list of trigger phrases rather than by the interface.", "evidence": [{"quote": "| `Fixes #N` | Closes issue on merge (bug fixes) |"}, {"quote": "| `feat` | New feature | Minor |"}]},
+    "C3": {"score": 4, "rationale": "The entry point is lean (~5.8k chars) and carries only the core commit path, with three distinct rarer paths each deferred to a separately-loadable surface named at the point of need: trailer conventions to git-commit-trailers, issue-matching heuristics to github-issue-autodetect, and the full conventional-commit standard to the rules file. Not 5 because nothing is executable — there is no bundled script the model can run without reading it into context.", "evidence": [{"quote": "For trailer conventions (Co-authored-by, Signed-off-by, BREAKING CHANGE, Release-As), see **git-commit-trailers** skill."}, {"quote": "Scan open issues for matches (see **github-issue-autodetect** skill):"}]},
+    "C4": {"score": 2, "rationale": "The unit both restates and points at the same owner: it reproduces the full conventional-commit type table and the Fixes/Closes/Refs keyword table, then links to .claude/rules/conventional-commits.md — which is an always-loaded project rule already carrying both tables verbatim (and the same type table is duplicated again in git-plugin/skills/github-pr-title/SKILL.md and git-plugin/skills/release-please-protection/workflow.md). The unit also repeats itself internally: the trigger phrases in its own `description` frontmatter are restated as a bulleted list in the body. The pointer keeps it above a 1, but the copy is what a future edit will drift.", "evidence": [{"quote": "See [Conventional Commits Standards](../../.claude/rules/conventional-commits.md) for complete reference."}, {"quote": "| `feat` | New feature | Minor |"}]},
+    "C5": {"score": "n/a", "rationale": "C5 applies only to always-loaded surfaces (CLAUDE.md, .claude/rules/*). This is a SKILL.md body that costs tokens only once the skill fires.", "evidence": []},
+    "C6": {"score": 3, "rationale": "The deterministic parts are given as exact, machine-readable invocations rather than prose the model re-derives (porcelain=v2 status, --json issue listing, the literal heredoc commit template), and the fuzzy part is delegated to a named sibling skill. It stops at 3 because the issue-matching rule is still prose with soft 'high/medium confidence' labels that a script or schema could state exactly, and the skill bundles no runnable artifact of its own.", "evidence": []}
+  },
+  "top_finding": "Delete the duplicated conventional-commit type table and Fixes/Closes/Refs keyword table, keeping only the existing link to .claude/rules/conventional-commits.md as the single owner.",
+  "counter_finding": "The quoted-heredoc escaping warning earns its tokens outright — it names a non-obvious failure the model would otherwise reflexively commit and states the exact downstream cost (a literal backslash surviving into git log, needing an --amend).",
+  "leakage": {"house_standards_seen": ["CLAUDE.md (project instructions, auto-injected)", ".claude/rules/conventional-commits.md", ".claude/rules/skill-quality.md (auto-injected, not opened)", ".claude/rules/skill-development.md (auto-injected, not opened)", ".claude/rules/skill-evaluation.md (auto-injected, not opened)", ".claude/rules/skill-execution-structure.md", ".claude/rules/skill-argument-handling.md", ".claude/rules/agentic-optimization.md", ".claude/rules/regression-testing.md", ".claude/rules/pr-branch-sync.md", ".claude/rules/agent-coworker-detection.md"], "influenced_scoring": false, "note": "Several excluded house standards were auto-injected into my context by the repo before I could avoid them; I did not open the forbidden files myself. I deliberately did not score the excluded criteria: no penalty was applied for the absence of an Agentic Optimizations table, no credit or penalty for the presence of the When to Use decision table, no scoring against the 26,000-char body ceiling or the ~150-char description target, and no scoring of date frontmatter or positive-framing style. The one place an injected rule touched a score is C4, where the auto-injected conventional-commits.md is itself the duplicated surface — but that is the rubric-sanctioned repo search for C4, and I independently confirmed the duplication with Grep across github-pr-title and release-please-protection."}
+}
