+++ docs/benchmarks/2026-07-context-engineering/judgments/test-tier-selection_j1.json
@@ -1,0 +1,16 @@
+{
+  "unit": "testing-plugin/skills/test-tier-selection/SKILL.md",
+  "unit_kind": "skill",
+  "judge_id": "j1",
+  "dimensions": {
+    "C1": {"score": 2, "rationale": "The core of the skill is a fixed change-type-to-tier matrix of unconditional Required/Skip/Optional verdicts with no stated rationale, over exactly the decision a capable model already makes by default; none of the constraints concerns safety, irreversibility, or an automation contract, and directives like 'All tiers must pass' / 'Fix immediately' fence defaults rather than delegate them. It escapes a 1 only because the Escalation Signals section frames some guidance as conditions rather than prohibitions.", "evidence": [{"quote": "| Bug fix (isolated) | Required | Optional | Skip |"}, {"quote": "2. All tiers must pass"}]},
+    "C2": {"score": 3, "rationale": "The skill is non-invocable and takes no arguments, so its interface is the vocabulary it declares: three tiers with scope/duration/when, and a decision matrix whose cells are drawn from an enumerated set (Required/Skip/Optional) — alternatives are stated at the point of declaration rather than taught by example. It stops short of higher because the four Quick Reference blocks are worked examples that restate the matrix instead of covering genuine ambiguity.", "evidence": []},
+    "C3": {"score": 3, "rationale": "The file is short and carries only the core selection path, deferring all execution to separately-loadable units it names (/test:quick, /test:full, /test:consult, and three agents) rather than inlining their content. It is not higher because it owns no separately-loadable reference of its own and inlines a redundant restatement block plus an Activation Triggers section that duplicates the frontmatter description.", "evidence": []},
+    "C4": {"score": 2, "rationale": "The tier-to-timing mapping is maintained three times inside this one file (Test Tier Definitions' 'When to Run' column, the Commands by Tier comments, and the Quick Reference headings), and a fourth copy lives in testing-plugin/README.md lines 38-44 ('Tiered Test Execution'), which has already drifted — the README maps Integration to `/test:full` while this file maps it to `/test:full --coverage`. Activation Triggers is a fifth restatement, of the frontmatter description.", "evidence": [{"quote": "# Tier 2: Integration (feature completion)"}, {"quote": "### After Feature Completion"}]},
+    "C5": {"score": "n/a", "rationale": "C5 applies only to always-loaded surfaces (CLAUDE.md, .claude/rules/*). This is a SKILL.md body that costs tokens only once the skill fires.", "evidence": []},
+    "C6": {"score": 2, "rationale": "The one deterministic step this skill actually owns — measuring the change's scope — is left entirely as prose criteria for the model to re-derive each run, even though it is pure counting over `git diff` and the frontmatter already grants Bash/Glob/Grep that the body never uses. It escapes a 1 because the Commands by Tier block does point at real runnable commands and is explicit that they are to be run.", "evidence": [{"quote": "- Changes span multiple files"}, {"quote": "allowed-tools: Bash, Read, Glob, Grep"}]}
+  },
+  "top_finding": "Replace the prose escalation criteria with a runnable scope-detection command (e.g. `git diff --name-only`/`--stat` against the base) so the tier decision is grounded in the actual change instead of re-derived from a static matrix each run.",
+  "counter_finding": "The Escalation Signals section earns its tokens: it names the underlying conditions (multi-file spans, service boundaries, auth changes, critical paths) that let the model generalise to change types the matrix never enumerates.",
+  "leakage": {"house_standards_seen": ["CLAUDE.md (project instructions, auto-injected)", ".claude/rules/skill-quality.md (auto-injected despite being on the do-not-read list)", ".claude/rules/skill-development.md (auto-injected, same)", ".claude/rules/skill-evaluation.md (auto-injected, same)", ".claude/rules/agentic-optimization.md", ".claude/rules/skill-execution-structure.md", ".claude/rules/skill-argument-handling.md", ".claude/rules/skill-consolidation.md", ".claude/rules/skill-naming.md", ".claude/rules/skill-fork-context.md", ".claude/rules/agentic-permissions.md", ".claude/rules/auto-mode.md", ".claude/rules/regression-testing.md", ".claude/rules/sandbox-guidance.md", ".claude/rules/version-pinning.md", ".claude/rules/offload-to-deterministic-substrate.md", "~12 other .claude/rules/*.md files"], "influenced_scoring": false, "note": "The repo auto-injected three files I was instructed not to read (skill-quality.md, skill-development.md, skill-evaluation.md) into my context before I could act; I did not open them deliberately and did not score against them. Concretely, I ignored every item on the rubric's exclusion list: the unit HAS the house-mandated '## When to Use This Skill' table and LACKS the house-mandated '## Agentic Optimizations' table, and I scored neither; I did not consider body char count, description length, the REFERENCE.md filename, date frontmatter, model policy, or positive-framing style. C3's reasoning is about disclosure structure (the rubric's own anchor language), not about the absence of a file with a particular house name. The one place a house rule and the rubric coincide is C6/offload-to-deterministic-substrate — I reached that finding from the rubric's 'rich references over prose specs' anchor and the unit's own unused `allowed-tools`, and would have scored it identically without the injected rule."}
+}
